developing encryption

+++ src/main/java/encryptor/encryptionInterface.java
@@ -2,8 +2,8 @@ package encryptor;
 
 public interface encryptionInterface {
 
-    Object[] Split(byte[] array_to_split);
-     byte[] ShiftLeft(byte[] array_to_shift, int cycles);
-    byte[] Merge(byte[] array_to_merge1, byte[] array_to_merge2);
+    Object[] Split(int[] array_to_split);
+     int[] ShiftLeft(int[] array_to_shift, int cycles);
+    int[] Merge(int[] array_to_merge1, int[] array_to_merge2);
 
 }
